Move key derivation into helper
